[#176175273]
Did some refactoring to make building the capitalis tool (bittensor cli) easier
